+++ changelog.md
@@ -14,3 +14,5 @@ This file is updated nearly every commit and copied to GH release changelog. Its
 * You can now import nougaro files from current folder and sub-folders.
 * Fixed an old bug with debug:
   * When you activate the debug mode from the shell, you no longer need to restart it for errors to print their origin file.
+* improve `__gpl__` builtin func on BSD:
+  * now can take any command as text editor